requirement.txt
torch
transformers
Pillow
faiss-cpu
numpy
matplotlib
tqdm
ultralytics
streamlit
opencv-python
scikit-learn
scipy